auth/requirements.txt
# Import the internal layer first
-r requirements-internal.txt

# Add Public Libs (External)
flask==3.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
pyjwt[crypto]==2.8.0 
cryptography>=41.0.0
gunicorn==21.2.0